refactor: :bug: Refactor Tests
