+++ tests/integration/routes/__init__.py
@@ -1,0 +1,4 @@
+from ..utils.factory import add_to_db
+
+
+add_to_db()
